skills/pyproject.toml
[project]
name = "skill-scripts"
version = "0.2.0"
description = "Shared Python dependencies for infrastructure skills"
requires-python = ">=3.12"
dependencies = [
    "tree-sitter>=0.23,<1.0",
    "tree-sitter-sql>=0.3,<1.0",
    "tree-sitter-python>=0.23,<1.0",
    "tree-sitter-typescript>=0.23,<1.0",
    "hvac>=2.1.0",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = [
    "worktree/scripts/tests",
    "coordination-bridge/scripts/tests",
    "validate-packages/scripts/tests",
    "validate-flows/scripts/tests",
    "refresh-architecture/scripts/tests",
    "bao-vault/scripts/tests",
]

[tool.ruff]
target-version = "py312"
line-length = 100
